Fix some scheduler 64 bit assertion failures

+++ xSchedule/PlayList/PlayListStep.cpp
@@ -570,11 +570,11 @@ std::string PlayListStep::FormatTime(size_t timems, bool ms) const
 {
     if (ms)
     {
-        return wxString::Format(wxT("%i:%02i.%03i"), timems / 60000, (timems % 60000) / 1000, timems % 1000).ToStdString();
+        return wxString::Format(wxT("%i:%02i.%03i"), (int)(timems / 60000), (int)((timems % 60000) / 1000), (int)(timems % 1000)).ToStdString();
     }
     else
     {
-        return wxString::Format(wxT("%i:%02i"), timems / 60000, (timems % 60000) / 1000).ToStdString();
+        return wxString::Format(wxT("%i:%02i"), (int)(timems / 60000), (int)((timems % 60000) / 1000)).ToStdString();
     }
 }
 
@@ -584,7 +584,7 @@ std::string PlayListStep::GetStatus(bool ms) const
 
     if (GetFrameMS() > 0)
     {
-        fps = wxString::Format(wxT("%i"), 1000 / GetFrameMS()).ToStdString();
+        fps = wxString::Format(wxT("%i"), (int)(1000 / GetFrameMS())).ToStdString();
     }
     
     return "Time: " + FormatTime(GetPosition(), ms) + " Left: " + FormatTime(GetLengthMS() - GetPosition(), ms) + " FPS: " + fps;
